in progress

+++ jpo-ode-plugins/src/test/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetB11BuilderTest.java
@@ -13,7 +13,7 @@ public class OffsetB11BuilderTest {
 
    @Test
    public void testConversion() {
-      int expectedValue = 312;
+      Long expectedValue = 312L;
       BigDecimal testInput = BigDecimal.valueOf(3.124);
 
       assertEquals(expectedValue,OffsetB11Builder.offsetB11(testInput));
